question-paper-analyzer/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pdfplumber==0.10.3
python-docx==0.8.11
Pillow==10.0.0
pytesseract==0.3.10
python-dotenv==1.0.1
google-genai==1.65.0
